Add files via upload
This is a simple google homepage for a non-signed-in account. 
The three; google logo, search icon(for which I could not easily find one with a transparent background) and the voice search icons are all images. 
The page is responsive and the search space is same width; 500px for both minimized and full screen just as in the real google page. This however changes when using smaller screens. i.e this is made possible by media queries which I will later add when making a signed-in account google homepage.
Feel free to share comments and insights.