Add files via upload
This CSS/HTML file designs a calculator; the outstanding features are grid columns for which I used a 1/3 ratio for the last column and a general 4 column ratio, 1fr. Another feature is the box-shadow, font type, background colors and padding. The margin was set to 0, auto. Feel free to try it out